build: Updating dependencies
composer:
* mediawiki/mediawiki-codesniffer: 49.0.0 → 50.0.0
* mediawiki/mediawiki-phan-config: 0.18.0 → 0.20.0

npm:
* flatted: 3.3.3 → 3.4.1
  * https://github.com/advisories/GHSA-25h7-pfq9-p65f

+++ includes/MapSourcesTransform.php
@@ -315,7 +315,6 @@ class MapSourcesTransform {
 	 */
 	protected function getUTM( &$utmArray ) {
 		$this->getUTMZone( $utmArray );
-		// @phan-suppress-next-line PhanTypePossiblyInvalidDimOffset
 		return $this->getTM( 0.0, $utmArray['centralMeridian'], $utmArray, $this->ellWGS84 );
 	}
 
@@ -344,10 +343,8 @@ class MapSourcesTransform {
 		}
 
 		# fix by Roger W Haworth
-		// @phan-suppress-next-line PhanTypePossiblyInvalidDimOffset
-		$gridX = floor( $utmArray['easting'] / 100000 );
-		// @phan-suppress-next-line PhanTypePossiblyInvalidDimOffset
-		$gridY = floor( $utmArray['northing'] / 100000 );
+		$gridX = (int)floor( $utmArray['easting'] / 100000 );
+		$gridY = (int)floor( $utmArray['northing'] / 100000 );
 
 		# outside area for OSGB36
 		if ( $gridX < 0 || $gridX > 6 || $gridY < 0 || $gridY > 12 ) {
@@ -361,9 +358,7 @@ class MapSourcesTransform {
 
 		$c1 = substr( $letters, ( 17 - intval( $gridY / 5 ) * 5 ) + intval( $gridX / 5 ), 1 );
 		$c2 = substr( $letters, ( 20 - ( $gridY % 5 ) * 5 ) + $gridX % 5, 1 );
-		// @phan-suppress-next-line PhanTypePossiblyInvalidDimOffset
 		$e = sprintf( "%05d", (int)$utmArray['easting'] % 100000 );
-		// @phan-suppress-next-line PhanTypePossiblyInvalidDimOffset
 		$n = sprintf( "%05d", (int)$utmArray['northing'] % 100000 );
 
 		$utmArray['error'] = 0;
